ASKMe-backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.1.1
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.13.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.4
llvmlite==0.44.0
lxml==5.3.1
MarkupSafe==2.1.5
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.3
numba==0.61.0
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
pillow==10.4.0
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
pyparsing==3.2.1
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
